Added tests and fixes

+++ tasks/app/services/tasks_crud.py
@@ -1,6 +1,7 @@
 from app.api.deps import db_session
 from app.models.Tasks import Task, TaskUpdate, TaskCreate
 from sqlmodel import select
+from sqlalchemy.orm import selectinload
 from app.core.exceptions.exceptions import TaskNotFoundException
 
 async def create_task(session: db_session, task_in: TaskCreate, owner_id: int):
@@ -11,14 +12,17 @@ async def create_task(session: db_session, task_in: TaskCreate, owner_id: int):
     return db_task
 
 async def fetch_task_by_id(session: db_session, task_id: int, owner_id: int):
-    result = await session.exec(select(Task).where(Task.id == task_id, Task.owner_id == owner_id))
+    result = await session.exec(
+        select(Task)
+        .options(selectinload(Task.category), selectinload(Task.streaks), selectinload(Task.sub_tasks))
+        .where(Task.id == task_id, Task.owner_id == owner_id))
     task = result.one_or_none()
     if not task:
         raise TaskNotFoundException()
     return task
 
 async def fetch_user_tasks(session: db_session, owner_id: int):
-    result = await session.exec(select(Task).where(Task.owner_id == owner_id))
+    result = await session.exec(select(Task).options(selectinload(Task.streaks)).where(Task.owner_id == owner_id))
     tasks = result.all()
     return tasks
 
@@ -43,7 +47,7 @@ async def remove_task(session: db_session, task_id: int, owner_id: int, delete_s
     if not db_task:
         raise TaskNotFoundException()
 
-    if delete_subtasks:
+    if delete_subtasks and db_task.sub_tasks: # Added check for sub_tasks
         for sub in db_task.sub_tasks:
             await session.delete(sub)
